# gen/http/graph/server/encode_decode.go
// Code generated by goa v3.24.1, DO NOT EDIT.
//
// graph HTTP server encoders and decoders
//
// Command:
// $ goa gen github.com/aditnikel/grapgraph/design

package server

import (
	"context"
	"errors"
	"io"
	"net/http"

	graph "github.com/aditnikel/grapgraph/gen/graph"
	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodeGetMetadataResponse returns an encoder for responses returned by the
// graph get_metadata endpoint.
func EncodeGetMetadataResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*graph.MetadataResponse)
		enc := encoder(ctx, w)
		body := NewGetMetadataResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// EncodePostSubgraphResponse returns an encoder for responses returned by the
// graph post_subgraph endpoint.
func EncodePostSubgraphResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*graph.SubgraphResponse)
		enc := encoder(ctx, w)
		body := NewPostSubgraphResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodePostSubgraphRequest returns a decoder for requests sent to the graph
// post_subgraph endpoint.
func DecodePostSubgraphRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*graph.SubgraphRequest, error) {
	return func(r *http.Request) (*graph.SubgraphRequest, error) {
		var (
			body PostSubgraphRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if errors.Is(err, io.EOF) {
				return nil, goa.MissingPayloadError()
			}
			var gerr *goa.ServiceError
			if errors.As(err, &gerr) {
				return nil, gerr
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		err = ValidatePostSubgraphRequestBody(&body)
		if err != nil {
			return nil, err
		}
		payload := NewPostSubgraphSubgraphRequest(&body)

		return payload, nil
	}
}

// EncodePostSubgraphError returns an encoder for errors returned by the
// post_subgraph graph endpoint.
func EncodePostSubgraphError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "bad_request":
			var res graph.BadRequest
			errors.As(v, &res)
			enc := encoder(ctx, w)
			body := res
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusBadRequest)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// marshalGraphGraphNodeToGraphNodeResponseBody builds a value of type
// *GraphNodeResponseBody from a value of type *graph.GraphNode.
func marshalGraphGraphNodeToGraphNodeResponseBody(v *graph.GraphNode) *GraphNodeResponseBody {
	res := &GraphNodeResponseBody{
		ID:    v.ID,
		Type:  v.Type,
		Key:   v.Key,
		Label: v.Label,
	}
	if v.Props != nil {
		res.Props = make(map[string]any, len(v.Props))
		for key, val := range v.Props {
			tk := key
			tv := val
			res.Props[tk] = tv
		}
	}

	return res
}

// marshalGraphGraphEdgeToGraphEdgeResponseBody builds a value of type
// *GraphEdgeResponseBody from a value of type *graph.GraphEdge.
func marshalGraphGraphEdgeToGraphEdgeResponseBody(v *graph.GraphEdge) *GraphEdgeResponseBody {
	res := &GraphEdgeResponseBody{
		ID:       v.ID,
		Type:     v.Type,
		From:     v.From,
		To:       v.To,
		Directed: v.Directed,
	}
	if v.Metrics != nil {
		res.Metrics = make(map[string]any, len(v.Metrics))
		for key, val := range v.Metrics {
			tk := key
			tv := val
			res.Metrics[tk] = tv
		}
	}

	return res
}