# gen/http/ingest/server/encode_decode.go
// Code generated by goa v3.24.1, DO NOT EDIT.
//
// ingest HTTP server encoders and decoders
//
// Command:
// $ goa gen github.com/aditnikel/grapgraph/design

package server

import (
	"context"
	"errors"
	"io"
	"net/http"

	ingest "github.com/aditnikel/grapgraph/gen/ingest"
	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodePostEventResponse returns an encoder for responses returned by the
// ingest post_event endpoint.
func EncodePostEventResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*ingest.IngestResponse)
		enc := encoder(ctx, w)
		body := NewPostEventResponseBody(res)
		w.WriteHeader(http.StatusAccepted)
		return enc.Encode(body)
	}
}

// DecodePostEventRequest returns a decoder for requests sent to the ingest
// post_event endpoint.
func DecodePostEventRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*ingest.CustomerEvent, error) {
	return func(r *http.Request) (*ingest.CustomerEvent, error) {
		var (
			body PostEventRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if errors.Is(err, io.EOF) {
				return nil, goa.MissingPayloadError()
			}
			var gerr *goa.ServiceError
			if errors.As(err, &gerr) {
				return nil, gerr
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		err = ValidatePostEventRequestBody(&body)
		if err != nil {
			return nil, err
		}
		payload := NewPostEventCustomerEvent(&body)

		return payload, nil
	}
}

// EncodePostEventError returns an encoder for errors returned by the
// post_event ingest endpoint.
func EncodePostEventError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "bad_request":
			var res ingest.BadRequest
			errors.As(v, &res)
			enc := encoder(ctx, w)
			body := res
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusBadRequest)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}