# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Grapgraph API
    description: High-performance fraud detection and financial network analysis platform built with Goa and FalkorDB.
    version: 0.0.1
host: localhost:8080
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /v1/graph/subgraph:
        post:
            tags:
                - graph
            summary: post_subgraph graph
            description: Extracts a surrounding subgraph for a specific root node using multi-hop analysis.
            operationId: graph#post_subgraph
            parameters:
                - name: post_subgraph_request_body
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/SubgraphRequest'
                    required:
                        - root
                        - time_window
                        - limit
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/SubgraphResponse'
                        required:
                            - version
                            - root
                            - nodes
                            - edges
                            - truncated
                "400":
                    description: Bad Request response.
                    schema:
                        type: string
            schemes:
                - http
    /v1/ingest/event:
        post:
            tags:
                - ingest
            summary: post_event ingest
            description: Accepts a new financial event (Payment, Login, etc.) and updates the relationship graph.
            operationId: ingest#post_event
            parameters:
                - name: post_event_request_body
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/CustomerEvent'
                    required:
                        - user_id
                        - event_type
                        - event_timestamp
            responses:
                "202":
                    description: Accepted response.
                    schema:
                        $ref: '#/definitions/IngestResponse'
                        required:
                            - accepted
                "400":
                    description: Bad Request response.
                    schema:
                        type: string
            schemes:
                - http
definitions:
    CustomerEvent:
        title: CustomerEvent
        type: object
        properties:
            device_id:
                type: string
                description: Unique hardware ID where the activity originated.
                example: d_888
            event_timestamp:
                description: Timestamp of the activity (RFC3339 string or Epoch MS).
                example: "2024-03-20T10:00:00Z"
            event_type:
                type: string
                description: The type of event (PAYMENT, LOGIN, WITHDRAWAL, etc).
                example: PAYMENT
            exchange:
                type: string
                description: Crypto exchange name if applicable.
                example: BINANCE
            ip_address:
                type: string
                description: Remote IP address (not stored directly in graph).
                example: 192.168.1.1
            issuing_bank:
                type: string
                description: The bank that issued the instrument.
                example: JP_MORGAN
            merchant_id_mpan:
                type: string
                description: Target merchant ID or card MPAN.
                example: m_777
            payment_method:
                type: string
                description: Method used (VISA, CRYPTO, etc).
                example: VISA
            total_transaction_amount:
                type: number
                description: Monetary value of the transaction.
                example: 150.5
                format: double
            user_id:
                type: string
                description: Unique identifier of the user (e.g. u_123).
                example: u_123
            wallet_address:
                type: string
                description: Blockchain wallet address if applicable.
                example: "0xabc123"
        example:
            device_id: d_888
            event_timestamp: "2024-03-20T10:00:00Z"
            event_type: PAYMENT
            exchange: BINANCE
            ip_address: 192.168.1.1
            issuing_bank: JP_MORGAN
            merchant_id_mpan: m_777
            payment_method: VISA
            total_transaction_amount: 150.5
            user_id: u_123
            wallet_address: "0xabc123"
        required:
            - user_id
            - event_type
            - event_timestamp
    GraphEdge:
        title: GraphEdge
        type: object
        properties:
            directed:
                type: boolean
                description: Whether the relationship has a specific flow direction.
                example: true
            from:
                type: string
                description: ID of the source node.
                example: USER:u_123
            id:
                type: string
                description: Unique ID for the specific relationship.
                example: e123
            metrics:
                type: object
                description: Statistical snapshots (count, amount, first_seen, etc).
                example:
                    Consequatur velit ea quibusdam odit rerum aliquam.: Enim aliquid accusamus accusantium ullam.
                    Nemo vel odio qui.: Et quo consectetur.
                additionalProperties: true
            to:
                type: string
                description: ID of the target node.
                example: MERCHANT:m_777
            type:
                type: string
                description: The type of connection (e.g. PAYMENT).
                example: PAYMENT
        description: A relationship between two entities with aggregated metrics.
        example:
            directed: true
            from: USER:u_123
            id: e123
            metrics:
                Natus ut autem possimus dolor sit iure.: Corrupti minima.
                Sunt rem voluptatem asperiores deserunt.: Fugit provident facere sapiente rerum.
            to: MERCHANT:m_777
            type: PAYMENT
        required:
            - id
            - type
            - from
            - to
            - directed
    GraphNode:
        title: GraphNode
        type: object
        properties:
            id:
                type: string
                description: Stable ID generated for visualization.
                example: USER:u_123
            key:
                type: string
                description: The domain-specific key (e.g. u_123).
                example: u_123
            label:
                type: string
                description: Human-friendly display name.
                example: User u_123
            props:
                type: object
                description: Additional key-value properties.
                example:
                    Et atque sit veniam quaerat quis.: Eum dolorum.
                additionalProperties: true
            type:
                type: string
                description: The category of the entity.
                example: USER
        description: A single entity (User, Merchant, Device) in the resulting subgraph.
        example:
            id: USER:u_123
            key: u_123
            label: User u_123
            props:
                Error doloremque quia quaerat.: Eum consequuntur dicta suscipit et hic veniam.
            type: USER
        required:
            - id
            - type
            - key
            - label
...
    IngestResponse:
        title: IngestResponse
        type: object
        properties:
            accepted:
                type: boolean
                description: Whether the event was successfully queued or processed.
                example: true
        example:
            accepted: true
        required:
            - accepted
...
    SubgraphRequest:
        title: SubgraphRequest
        type: object
        properties:
            edge_types:
                type: array
                items:
                    type: string
                    example: Vel et porro.
                description: Filter to only include these relationship types.
                example:
                    - PAYMENT
                    - LOGIN
            hops:
                type: integer
                description: Number of hops to traverse (1-3).
                default: 2
                example: 2
                format: int64
                minimum: 1
                maximum: 3
            limit:
                type: object
                properties:
                    max_edges:
                        type: integer
                        description: Maximum number of edges to return.
                        default: 200
                        example: 100
                        format: int64
                    max_nodes:
                        type: integer
                        description: Maximum number of nodes to return.
                        default: 100
                        example: 50
                        format: int64
                description: Resource budget for the response.
                example:
                    max_edges: 100
                    max_nodes: 50
                required:
                    - max_nodes
                    - max_edges
            min_event_count:
                type: integer
                description: Minimum number of aggregate events to include an edge.
                default: 1
                example: 2
                format: int64
            rank_neighbors_by:
                type: string
                description: Metric used to sort and truncate neighbor nodes.
                default: event_count_30d
                example: total_amount
            root:
                type: object
                properties:
                    key:
                        type: string
                        description: The unique key of the root node.
                        example: u_123
                    type:
                        type: string
                        description: Type of the root node (usually USER).
                        example: USER
                description: The starting node for the traversal.
                example:
                    key: u_123
                    type: USER
                required:
                    - type
                    - key
            time_window:
                type: object
                properties:
                    from:
                        type: string
                        description: Start of the window (RFC3339).
                        example: "2024-01-01T00:00:00Z"
                    to:
                        type: string
                        description: End of the window (RFC3339).
                        example: "2024-12-31T23:59:59Z"
                description: Optional time range to filter relationship metrics.
                example:
                    from: "2024-01-01T00:00:00Z"
                    to: "2024-12-31T23:59:59Z"
                required:
                    - from
                    - to
        example:
            edge_types:
                - PAYMENT
                - LOGIN
            hops: 2
            limit:
                max_edges: 100
                max_nodes: 50
            min_event_count: 2
            rank_neighbors_by: total_amount
            root:
                key: u_123
                type: USER
            time_window:
                from: "2024-01-01T00:00:00Z"
                to: "2024-12-31T23:59:59Z"
        required:
            - root
            - time_window
            - limit
    SubgraphResponse:
        title: SubgraphResponse
        type: object
        properties:
            edges:
                type: array
                items:
                    $ref: '#/definitions/GraphEdge'
                description: List of all connections found.
                example:
                    - directed: true
                      from: USER:u_123
                      id: e123
                      metrics:
                        Accusantium repellat dolores quae.: Ratione dolor natus dolor maiores magnam.
                      to: MERCHANT:m_777
                      type: PAYMENT
                    - directed: true
                      from: USER:u_123
                      id: e123
                      metrics:
                        Accusantium repellat dolores quae.: Ratione dolor natus dolor maiores magnam.
                      to: MERCHANT:m_777
                      type: PAYMENT
                    - directed: true
                      from: USER:u_123
                      id: e123
                      metrics:
                        Accusantium repellat dolores quae.: Ratione dolor natus dolor maiores magnam.
                      to: MERCHANT:m_777
                      type: PAYMENT
                    - directed: true
                      from: USER:u_123
                      id: e123
                      metrics:
                        Accusantium repellat dolores quae.: Ratione dolor natus dolor maiores magnam.
                      to: MERCHANT:m_777
                      type: PAYMENT
            nodes:
                type: array
                items:
                    $ref: '#/definitions/GraphNode'
                description: List of all entities in the network.
                example:
                    - id: USER:u_123
                      key: u_123
                      label: User u_123
                      props:
                        Eius quis est assumenda excepturi quod.: Exercitationem tempore aspernatur.
                        Et et natus minus ea eligendi.: Non veniam vel rerum sed placeat natus.
                        Vel reprehenderit facere.: Esse doloremque occaecati enim est unde qui.
                      type: USER
                    - id: USER:u_123
                      key: u_123
                      label: User u_123
                      props:
                        Eius quis est assumenda excepturi quod.: Exercitationem tempore aspernatur.
                        Et et natus minus ea eligendi.: Non veniam vel rerum sed placeat natus.
                        Vel reprehenderit facere.: Esse doloremque occaecati enim est unde qui.
                      type: USER
                    - id: USER:u_123
                      key: u_123
                      label: User u_123
                      props:
                        Eius quis est assumenda excepturi quod.: Exercitationem tempore aspernatur.
                        Et et natus minus ea eligendi.: Non veniam vel rerum sed placeat natus.
                        Vel reprehenderit facere.: Esse doloremque occaecati enim est unde qui.
                      type: USER
                    - id: USER:u_123
                      key: u_123
                      label: User u_123
                      props:
                        Eius quis est assumenda excepturi quod.: Exercitationem tempore aspernatur.
                        Et et natus minus ea eligendi.: Non veniam vel rerum sed placeat natus.
                        Vel reprehenderit facere.: Esse doloremque occaecati enim est unde qui.
                      type: USER
            root:
                type: string
                description: The ID of the requested starting node.
                example: USER:u_123
            truncated:
                type: boolean
                description: Indicates if the result was clipped by performance budgets.
                example: false
            version:
                type: string
                description: Format version of the response.
                example: "1.0"
        example:
            edges:
                - directed: true
                  from: USER:u_123
                  id: e123
                  metrics:
                    Accusantium repellat dolores quae.: Ratione dolor natus dolor maiores magnam.
                  to: MERCHANT:m_777
                  type: PAYMENT
                - directed: true
                  from: USER:u_123
                  id: e123
                  metrics:
                    Accusantium repellat dolores quae.: Ratione dolor natus dolor maiores magnam.
                  to: MERCHANT:m_777
                  type: PAYMENT
                - directed: true
                  from: USER:u_123
                  id: e123
                  metrics:
                    Accusantium repellat dolores quae.: Ratione dolor natus dolor maiores magnam.
                  to: MERCHANT:m_777
                  type: PAYMENT
            nodes:
                - id: USER:u_123
                  key: u_123
                  label: User u_123
                  props:
                    Eius quis est assumenda excepturi quod.: Exercitationem tempore aspernatur.
                    Et et natus minus ea eligendi.: Non veniam vel rerum sed placeat natus.
                    Vel reprehenderit facere.: Esse doloremque occaecati enim est unde qui.
                  type: USER
                - id: USER:u_123
                  key: u_123
                  label: User u_123
                  props:
                    Eius quis est assumenda excepturi quod.: Exercitationem tempore aspernatur.
                    Et et natus minus ea eligendi.: Non veniam vel rerum sed placeat natus.
                    Vel reprehenderit facere.: Esse doloremque occaecati enim est unde qui.
                  type: USER
            root: USER:u_123
            truncated: false
            version: "1.0"
        required:
            - version
            - root
            - nodes
            - edges
            - truncated